Add support for IPVS forward method
Add annotation and environment variable parsing to allow specifying the
forwarding method for the underlying IPVS configuration.

This also fixes an existing bug where configmap entries could be
orphaned if the IP address changed.  This is done by persisting all
necessary service information in the annotation and generating the
entire config map from that each time.